+++ src/serial/huffman.h
@@ -22,6 +22,6 @@ void freeHuffmanAlphabet(char* huffmanAlphabet[]);
 ** Write a single huffman encoded bit to the output file chunk buffer.
 ** Returns true in case of success, false otherwise.
 */
-bool fwriteBitInBuffer(char bit, FILE* outputFile, char buffer[], int* nbits, int* nbytes, int* outputSize);
+bool fwriteBitInBuffer(char bit, FILE* outputFile, char buffer[], int* nbits, ull* nbytes, ull* outputSize);
 
 #endif
